+++ packages/test-utils/src/find-vue-components.js
@@ -47,15 +47,13 @@ function findAllFunctionalComponentsFromVnode (
 
 export function vmCtorMatchesName (vm: Component, name: string): boolean {
   return !!(
-    (vm.$vnode &&
-      vm.$vnode.componentOptions &&
-      vm.$vnode.componentOptions.Ctor.options.name === name) ||
-    (vm._vnode &&
+    name && (
+      (vm._vnode &&
       vm._vnode.functionalOptions &&
       vm._vnode.functionalOptions.name === name) ||
     (vm.$options && vm.$options.name === name) ||
     (vm.options && vm.options.name === name)
-  )
+    ))
 }
 
 export function vmCtorMatchesSelector (component: Component, selector: Object) {
@@ -106,7 +104,7 @@ export default function findVueComponents (
     )
   }
   const nameSelector =
-    typeof selector === 'function' ? selector.options.name : selector.name
+    typeof selector === 'function' ? selector.extendOptions.name : selector.name
   const components = root._isVue
     ? findAllVueComponentsFromVm(root)
     : findAllVueComponentsFromVnode(root)