+++ test/specs/wrapper/setSelected.spec.js
@@ -33,19 +33,6 @@ describeWithShallowAndMount('setSelected', mountingMethod => {
     expect(wrapper.text()).to.contain('selectA')
   })
 
-  it('throws error if wrapper does not contain element', () => {
-    const wrapper = mountingMethod({ template: '<div><p/></div>' })
-    const p = wrapper.find('p')
-    p.element = null
-
-    const fn = () => p.setSelected()
-    const message =
-      '[vue-test-utils]: cannot call wrapper.setSelected() on a wrapper without an element'
-    expect(fn)
-      .to.throw()
-      .with.property('message', message)
-  })
-
   it('throws error if element is radio', () => {
     const message =
       'wrapper.setSelected() cannot be called on a <input type="radio" /> element. Use wrapper.setChecked() instead'
@@ -62,6 +49,7 @@ describeWithShallowAndMount('setSelected', mountingMethod => {
     const message =
       'wrapper.setSelected() cannot be called on "text" inputs. Use wrapper.setValue() instead'
     shouldThrowErrorOnElement('input[type="text"]', message)
+    shouldThrowErrorOnElement('textarea', message)
   })
 
   it('throws error if element is not valid', () => {